requirements.txt
boto3==1.14.30
snowflake-connector-python==2.2.9
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.18